Added DefaultRedisCacheConfiguration and refactored existing config
This will allow users to specify default configuration values for all their Redis caches, e.g. a default serializer. Previously this serializer had to be specified for each cache individually which adds noise to the configuration file. This applies to all properties except charset which is inherited from the application context.

Both DefaultRedisCacheConfiguration and RedisCacheConfiguration extend AbstractRedisCacheConfiguration to get optimal overlap of configuration without duplicate code. The only difference is that RedisCacheConfiguration provides the cache name, as AbstractRedisCacheConfiguration doesn't know it.