Handling exceptions caused by value serializers
If serialization and deserialization exceptions aren't handled, result CompletableFuture's will never be completed and cause hangs wherever they are accessed.

This fix wraps all serialization and deserialization in try-catch and completes the CompletableFuture's exceptionally after catching an exception.

This is only necessary for the async cache, the sync cache doesn't run into this problem and doesn't swallow such exceptions.